星評価を0.5刻みで付けられるようにする
DBのrating列をsmallint（1〜5の整数のみ）からnumeric(2,1)に変更し、
0.5刻みであることを検証する制約に置き換え。Flutter側のGameLog.ratingも
int?からdouble?に変更し、StarRatingウィジェットのallowHalfRatingを有効化。
follow_feedビューはratingに依存しているため型変更時に一旦作り直している。

+++ test/features/summary/period_summary_test.dart
@@ -7,7 +7,7 @@ GameLogWithGame _entry({
   required int id,
   required GameLogStatus status,
   required DateTime createdAt,
-  int? rating,
+  double? rating,
   List<String> genres = const [],
 }) {
   return GameLogWithGame(
